feat: UI for loading and unlocking databases
