+++ www/src/omnikee.ts
@@ -4,7 +4,12 @@ import {type OTPResponse, type DatabaseOverview, type Entry} from 'omnikee-wasm'
 
 export interface OmniKee {
   listDatabases(): Promise<DatabaseOverview[]>;
-  loadDatabase(data: Uint8Array, password: string | null, keyfile: Uint8Array | null): Promise<DatabaseOverview>;
+  loadDemo(): Promise<DatabaseOverview>;
+  loadDatabase(): Promise<DatabaseOverview>;
+
+  unlockDatabase(databaseIdx: number, password: string | null, keyfile: Uint8Array | null): Promise<DatabaseOverview>,
+  lockDatabase(databaseIdx: number): Promise<DatabaseOverview>,
+
   closeDatabase(databaseIdx: number): Promise<void>,
 
   listEntries(databaseIdx: number, groupUuid: string): Promise<Entry[]>,
@@ -27,9 +32,46 @@ if (process.env.TAURI_ENV_PLATFORM === 'web') {
 
   const state = ok.AppState.new()
 
+
+  function promptFile(): Promise<File> {
+
+    return new Promise((resolve, reject) => {
+
+      const $input = document.createElement('input')
+      $input.type = 'file'
+
+      if (!$input) {throw Error("could not create input")}
+
+      $input.addEventListener('change', (event) => {
+        const files = (event.target as HTMLInputElement).files
+
+        if (files && files[0]) {
+          resolve(files[0])
+        } else {
+          reject(new Error("no file selected"))
+        }
+      })
+
+      $input.click();
+    })
+  }
+
+
   handle = {
     listDatabases() {return Promise.resolve(state.list_databases())},
-    loadDatabase(data, password, keyfile) {return Promise.resolve(state.load_database_buffer(data, password, keyfile))},
+
+    loadDemo() {return Promise.resolve(state.load_demo())},
+
+    async loadDatabase() {
+      const file = await promptFile()
+      const data = new Uint8Array(await file.arrayBuffer())
+
+      return Promise.resolve(state.load_database_buffer(file.name, data))
+    },
+
+    unlockDatabase(databaseIdx, password, keyfile) {return Promise.resolve(state.unlock_database(databaseIdx, password, keyfile))},
+    lockDatabase(databaseIdx) {return Promise.resolve(state.lock_database(databaseIdx))},
+
     closeDatabase(databaseIdx) {return Promise.resolve(state.close_database(databaseIdx))},
 
     listEntries(databaseIdx, groupUuid) {return Promise.resolve(state.list_entries(databaseIdx, groupUuid))},
@@ -59,7 +101,16 @@ if (process.env.TAURI_ENV_PLATFORM === 'web') {
 
   handle = {
     async listDatabases() {return await invoke('list_databases')},
-    async loadDatabase(data, password, keyfile) {return await invoke('load_database_buffer', {data, password, keyfile})},
+
+    async loadDemo() {return await invoke('load_demo')},
+
+    async loadDatabase() {
+      // file picking is done from the Tauri backend
+      return await invoke('load_database')
+    },
+
+    async unlockDatabase(databaseIdx, password, keyfile) {return await invoke('unlock_database', {databaseIdx, password, keyfile})},
+    async lockDatabase(databaseIdx) {return await invoke('lock_database', {databaseIdx})},
     async closeDatabase(databaseIdx) {return await invoke('close_database', {databaseIdx})},
 
     async listEntries(databaseIdx, groupUuid) {return await invoke<Entry[]>('list_entries', {databaseIdx, groupUuid})},
